examples/python/requirements.txt
grpcio==1.76.0
protobuf==7.34.1

